requirements.txt
tensorflow>=2.8.0
numpy>=1.21.0
matplotlib>=3.5.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
requests>=2.25.0
Pillow>=8.0.0